requirements.txt
Flask>=3.0,<4.0
Flask-SQLAlchemy>=3.1,<4.0
Flask-WTF>=1.2,<2.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
python-pptx>=0.6.23,<1.0
Pillow>=10.0,<12.0
